requirements.txt
dvc
pymongo
python-dotenv